+++ src/briefy/choreographer/actions/slack/leica/order.py
@@ -1,7 +1,7 @@
 """Slack actions for Orders."""
 from briefy.choreographer.actions.slack import ISlack
 from briefy.choreographer.actions.slack import Slack
-from briefy.choreographer.events.leica import order as events
+from briefy.choreographer.events.leica.order import order as events
 from zope.component import adapter
 from zope.interface import implementer