+++ src/briefy/choreographer/subscribers/logger.py
@@ -1,5 +1,6 @@
 """Logger handler."""
 from briefy.choreographer.events import InternalEvent
+from briefy.choreographer.utils import get_actions_for_event
 from briefy.common.utils.transformers import json_dumps
 
 import logging
@@ -11,14 +12,13 @@ logger = logging.getLogger(__name__)
 def handler(event: InternalEvent):
     """Catch all handler that logs the event.
 
-    :param event: Event
+    :param event: InternalEvent instance
     """
+    actions = get_actions_for_event(event)
+    actions_info = [cls.action_info for cls in actions]
+    total_actions = len(actions)
     logger.info(
-        'Event: {name}, Entity: {entity}, GUID: {guid}'.format(
-            name=event.event_name,
-            entity=event.entity,
-            guid=event.guid
-        ),
+        f'Event {event.event_name} ({event.guid}) with {total_actions:02d} actions',
         extra={
             'choreographer': {
                 'event_name': event.event_name,
@@ -27,7 +27,9 @@ def handler(event: InternalEvent):
                 'entity': event.entity,
                 'actor': event.actor,
                 'request_id': event.request_id,
-                'data_payload': json_dumps(event.data)
+                'data_payload': json_dumps(event.data),
+                'total_actions': total_actions,
+                'registered_actions': actions_info
             }
         }
     )