+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+2.0.20 (2017-08-22)
+-------------------
+
+* Card #552: LeadOrder events not being sent (ericof).
+* Improve logging of discarded events (ericof).
+
 2.0.19 (2017-08-07)
 -------------------